+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Symlink Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [Initial Version] - 2025-03-06
+- Added Create Symlink command